requirement.txt
fastapi[all]
pydantic-settings
requests
langchain